+++ bruker_control/bruker_control.py
@@ -6,7 +6,7 @@
 # https://github.com/PowerBroker2/pySerialTransfer
 # Genie Nano manufactured by Teledyne DALSA
 
-__version__ = "1.0.0"
+__version__ = "1.6.0"
 
 ###############################################################################
 # Import Packages