+++ bruker_control/config_utils.py
@@ -17,10 +17,15 @@ from pathlib import Path
 # Import datetime for folder naming
 from datetime import datetime
 
+# Import YAML for gathering metadata about project
+from ruamel.yaml import YAML
+
 # Template configuration directories are within project directories.  The snlkt
 # server housing these directories is mounted to the X: volume on the machine
 # BRUKER.
+# TODO: This obviously should be the same thing
 base_template_config_dir = Path("X:/")
+server_basepath = "X:/"
 
 # Experimental configuration directories are in the Raw Data volume on the
 # machine BRUKER which is mounted to E:. This is where configs will be written
@@ -104,7 +109,6 @@ def read_config(config_file_path: Path) -> dict:
 
     Returns:
         Dictionary of contents inside the configuration .json file
-
     """
 
     with open(config_file_path, 'r') as inFile:
@@ -170,18 +174,18 @@ def write_experiment_config(config_template: dict, experiment_arrays: list,
 
     # Assign trialArray key to trialArray data.  The 0th index of the list is
     # always the the trialArray
-    config_template["trialArray"] = experiment_arrays[0]
+    config_template["beh_metadata"]["trialArray"] = experiment_arrays[0]
 
     # Assign ITIArray key to ITIArray data.  The 1st index of the list is
     # always the ITIArray
-    config_template["ITIArray"] = experiment_arrays[1]
+    config_template["beh_metadata"]["ITIArray"] = experiment_arrays[1]
 
     # Assign toneArray key the toneArray data.  The 2nd index of the list is
     # always the toneArray.
-    config_template["toneArray"] = experiment_arrays[2]
+    config_template["beh_metadata"]["toneArray"] = experiment_arrays[2]
 
     # Assign dropped_frames key the dropped_frames data.
-    config_template["dropped_frames"] = dropped_frames
+    config_template["beh_metadata"]["dropped_frames"] = dropped_frames
 
     # # Write the new configuration file
     with open(config_fullpath, 'w') as outFile:
@@ -215,7 +219,124 @@ def get_arduino_metadata(config_template: dict) -> dict:
 
     # Generate Dictionary of relevant Arduino metadata
     arduino_metadata = {key: value for (key, value) in
-                        config_template["metadata"].items() if
+                        config_template["beh_metadata"].items() if
                         key in arduino_metadata_keys}
 
     return arduino_metadata
+
+
+def get_zstack_metadata(config_template: dict) -> dict:
+    """
+    Grabs metadata relevant for generating Z-stacks
+
+    Parses template configuration supplied by the user and grabs only the
+    metadata that is relevant for Prairie View executing a Z-stack.
+
+    Args:
+        config_template:
+            Configuration template value dictionary gathered from team's
+            configuration .json file.
+
+    Returns:
+        zstack_metadata
+            Dictionary of relevant Prairie View Z-Stack for experiment
+    """
+
+    zstack_metadata = config_template["zstack_metadata"]
+
+    return zstack_metadata
+
+
+def get_subject_metadata(team: str, subject_id: str) -> dict:
+    """
+    Parses imaging subject's .yml metadata file for NWB fields
+
+    Locates and then uses ruamel.yaml to parse the metadata fields with safe
+    loading. Gathers yaml data and places it into a dictionary for use later
+    in the NWB file.
+
+    Args:
+        team:
+            Team value from metadata_args["team"]
+        subject_id:
+            Subject ID from metadata_args["subject"]
+
+    Returns:
+        subject_metadata
+    """
+
+    # Define YAML object parser with safe loading
+    yaml = YAML(typ='safe')
+
+    # Construct the base path for the subject's YAML file
+    base_yaml_path = Path(server_basepath + team + "/animal_metadata/")
+
+    animal_glob = [subject for subject in
+                   base_yaml_path.glob(f"{subject_id}.yml")]
+
+    # TODO: Raise warning here if there's more than one animal presented in
+    # this glob
+    subject_metadata = yaml.load(animal_glob[0])
+
+    return subject_metadata
+
+
+def get_surgery_metadata(subject_metadata: dict) -> dict:
+    """
+    Grabs surgery information from subject's metadata file.
+
+    Implant information and virus information for stimulation and recording
+    is contained within the subject's metadata file.  This information is used
+    for filenaming purposes particularly in the Z-Stack files as well as the
+    NWB files' metadata sections.
+
+    Args:
+        subject_metadata:
+            Metadata obtained from get_subject_metadata()
+
+    Returns:
+        surgery_metadata
+    """
+
+    surgery_metadata = subject_metadata["surgery"]
+
+    surgery_metadata = surgery_metadata[next(iter(surgery_metadata))]
+
+    return surgery_metadata
+
+
+def get_project_metadata(team: str, subject_id: str):
+    """
+    Grabs and parses project metadata yml file for NWB file generation.
+
+    Each project has its own metadata associated with it that NWB uses in its
+    standard.  This function grabs the proper file and builds a dictionary that
+    is used when populating metadata later.
+
+    Args:
+        team:
+            Team value from metadata_args["team"]
+        subject_id:
+            Subject ID from metadata_args["subject"]
+
+    Returns:
+        project_metadata
+    """
+
+    # Define YAML object parser with safe loading
+    yaml = YAML(typ='safe')
+
+    # Construct the base path for the project's YAML file
+    base_yaml_path = server_basepath + team + "/2p_template_configs/"
+
+    # Until teams and studies/projects are implemented across all directories,
+    # this if/else will have to do
+    if "LH" in subject_id:
+        project_yaml_path = Path(base_yaml_path + "nwb_lh_base.yml")
+    else:
+        project_yaml_path = base_yaml_path + "nwb_cs_base.yml"
+
+    # Load the project metadata into a dictionary
+    project_metadata = yaml.load(project_yaml_path)
+
+    return project_metadata